requirements.txt
# LangChain + LangGraph
langchain==1.2.13
langchain-core==1.2.23
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-text-splitters==1.1.1
langchain-classic==1.0.3
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8

# LLM — instalar el paquete del proveedor que uses en LLM_MODEL
# LLM_MODEL=anthropic/...      → pip install langchain-anthropic
# LLM_MODEL=openai/...         → pip install langchain-openai
# LLM_MODEL=ollama/...         → pip install langchain-ollama  (+ ollama corriendo local)
# LLM_MODEL=google_genai/...   → pip install langchain-google-genai
# LLM_MODEL=groq/...           → pip install langchain-groq
langchain-anthropic==1.4.0  # proveedor por defecto
anthropic==0.86.0
# langchain-openai
# langchain-ollama
# langchain-google-genai
# langchain-groq

# Vector store
chromadb==1.5.5

# Embeddings locales
sentence-transformers==5.3.0

# Cache
redis==7.4.0

# Utils
python-dotenv==1.2.2
rich==14.3.3
pydantic==2.12.5
pydantic-settings==2.13.1

# Observabilidad (módulo 6)
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
